chore: rename bumpSpecVersion to specVersion
